refactor(server): extract Lovable prompt-pack → src/server/lovable.js
Pure move, zero behavior change. Pulls the entire Lovable integration
(Brand Intelligence Profile -> deterministic URL-encoded build prompt)
out of server.js into src/server/lovable.js — ~324 lines, 17 helpers +
4 consts.

Why: continues the server.js dismemberment. The Lovable suite is a
fully self-contained leaf — pure templating, no LLM/DB/network calls,
zero external symbol references — so it lifts out cleanly as one unit.

What:
- New src/server/lovable.js holds the LOVABLE_* consts and all
  lovable* helpers. Exports the 4 consts + the 13 functions the
  /api/forge/prompt-pack/lovable route handler consumes (builders,
  formatters, naming helpers). lovableTruncate / lovableSection /
  lovableProductTypeLabel stay module-private (only used internally).
- server.js imports that surface and keeps the route handler in place;
  the inline definitions are replaced with a breadcrumb comment.
- Code moved verbatim — no logic touched.

Test plan:
- node --check server.js -> OK
- npm run lint (no-undef gate) -> clean (confirms every symbol the
  handler references is imported)
- route-inventory guard -> PASS (snapshot unchanged — route untouched)
- vitest -> 55/55 pass (9 new lovable tests: uuid/app-type gates,
  hasData/safeJoin, voice/persona/geo formatters, both builders, stub +
  naming helpers)

Rollback: revert this commit; the suite was a pure copy, so reverting
restores the inline block exactly.
